feat: Add server id to replicated writes
Adds the server id to replicated writes. Changes type from u8 to u16 for more potential write servers. Also updates the server struct in Delorean to no longer have locks. Locking will happen higher up, likely with the server being arc_swap'd out on any configuration update.

+++ delorean_write_buffer/src/database.rs
@@ -199,7 +199,7 @@ pub enum Error {
     GenericQueryError { message: String, query: String },
 
     #[snafu(display("replicated write from writer {} missing payload", writer))]
-    MissingPayload { writer: u8 },
+    MissingPayload { writer: u16 },
 }
 
 impl From<crate::table::Error> for Error {
